test: keep the unwritable-home doctor check off Windows
